11103: ensure the index exists before attempting to drop it.

+++ services/api/db/migrate/20170216170823_no_cr_mounts_and_workflow_def_in_full_text_search_index.rb
@@ -22,7 +22,13 @@ class NoCrMountsAndWorkflowDefInFullTextSearchIndex < ActiveRecord::Migration
 
   def down
     fts_indexes.each do |t, i|
-      remove_index t.to_sym, :name => i
+      t.classify.constantize.reset_column_information
+      ActiveRecord::Base.connection.indexes(t).each do |idx|
+        if idx.name == i
+          remove_index t.to_sym, :name => i
+          break
+        end
+      end
     end
   end
 end
